Chore: move payload to service pkg
