+++ internal/pkg/service/responder.go
@@ -1,11 +1,9 @@
 package service
 
 import (
-	"cqrses/pkg/payload"
-
 	"github.com/labstack/echo/v4"
 )
 
 type Responder interface {
-	Response(c echo.Context, payload payload.DomainPayload) error
+	Response(c echo.Context, payload DomainPayload) error
 }
